Refactor how I use the password rules helper
